refactor: 컴포넌트 파일명 PascalCase 통일 (#29 - Step 3)
소문자로 시작하던 컴포넌트 파일명을 PascalCase로 통일

- mobileHeader.jsx → MobileHeader.jsx
- mobileTabBar.jsx → MobileTabBar.jsx
- mobileEventRow.jsx → MobileEventRow.jsx
