+++ src/components/main/mobile/feature/EVL/MobileEventRow.jsx
