+++ src/pages/main/CBL/CBLPage.jsx
@@ -1,6 +1,6 @@
 import { useDeviceStore } from "@/stores/deviceStore";
-import MobileHeader from "@/components/main/mobile/common/mobileHeader";
-import MobileTabBar from "@/components/main/mobile/common/mobileTabBar";
+import MobileHeader from "@/components/main/mobile/common/MobileHeader";
+import MobileTabBar from "@/components/main/mobile/common/MobileTabBar";
 import MobileFooter from "@/components/main/mobile/common/MobileFooter";
 import TabBar from "@/components/main/desktop/common/TabBar";
 import Footer from "@/components/main/desktop/common/Footer";
