+++ src/pages/main/EVD/EVDPage.jsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import TabBar from "@/components/main/desktop/common/TabBar";
 import Footer from "@/components/main/desktop/common/Footer";
-import MobileHeader from "@/components/main/mobile/common/mobileHeader";
-import MobileTabBar from "@/components/main/mobile/common/mobileTabBar";
+import MobileHeader from "@/components/main/mobile/common/MobileHeader";
+import MobileTabBar from "@/components/main/mobile/common/MobileTabBar";
 import MobileFooter from "@/components/main/mobile/common/MobileFooter";
 import EventDetail from "@/components/main/adaptive/feature/EVD/EventDetail";
 import { fetchEventDetail } from "@/api/articles";
