+++ src/pages/main/EVL/EVLPage.jsx
@@ -1,18 +1,18 @@
 import { useState, useEffect, useMemo } from "react";
-import MobileHeader from "@/components/main/mobile/common/mobileHeader";
+import MobileHeader from "@/components/main/mobile/common/MobileHeader";
 import { useNavigate } from "react-router-dom";
 import { getStatus } from "@/utils/statusUtil";
 import TabBar from "@/components/main/desktop/common/TabBar";
 import Footer from "@/components/main/desktop/common/Footer";
 import MiniCalendarSet from "@/components/main/desktop/common/MiniCalendarSet";
 import EventRow from "@/components/main/adaptive/feature/EVL/EventRow";
-import MobileEventRow from "@/components/main/mobile/feature/EVL/mobileEventRow";
+import MobileEventRow from "@/components/main/mobile/feature/EVL/MobileEventRow";
 import FilterBottomSheet from "@/components/main/adaptive/feature/EVL/FilterBottomSheet";
 import SearchBar from "@/components/main/adaptive/common/SearchBar";
 import { FiFilter } from "react-icons/fi";
 import { fetchEvents } from "@/api/articles";
 import { useDeviceStore } from "@/stores/deviceStore";
-import MobileTabBar from "@/components/main/mobile/common/mobileTabBar";
+import MobileTabBar from "@/components/main/mobile/common/MobileTabBar";
 import SectionTitle from "@/components/main/mobile/common/SectionTitle";
 import useEVLFilterStore from "@/stores/useEVLFilterStore";
 import MobileFooter from "@/components/main/mobile/common/MobileFooter";
