+++ src/pages/main/HOM/HOMPage.jsx
@@ -4,8 +4,8 @@ import ServiceLinkList from "@/components/main/desktop/common/ServiceLinkList";
 import CalendarSection from "@/components/main/adaptive/feature/HOM/CalendarSection";
 import HotEventList from "@/components/main/mobile/feature/HOM/HotEventList";
 import { useDeviceStore } from "@/stores/deviceStore";
-import MobileHeader from "@/components/main/mobile/common/mobileHeader";
-import MobileTabBar from "@/components/main/mobile/common/mobileTabBar";
+import MobileHeader from "@/components/main/mobile/common/MobileHeader";
+import MobileTabBar from "@/components/main/mobile/common/MobileTabBar";
 import MobileFooter from "@/components/main/mobile/common/MobileFooter";
 import { useEffect } from "react";
 import { useQueryClient } from "@tanstack/react-query";
