+++ src/pages/main/MYP/MYPage.jsx
@@ -4,8 +4,8 @@ import Footer from "@/components/main/desktop/common/Footer";
 import ProfileSection from "@/components/main/adaptive/feature/MYP/ProfileSection";
 import BookmarkSection from "@/components/main/adaptive/feature/MYP/BookmarkSection";
 import { useDeviceStore } from "@/stores/deviceStore";
-import MobileHeader from "@/components/main/mobile/common/mobileHeader";
-import MobileTabBar from "@/components/main/mobile/common/mobileTabBar";
+import MobileHeader from "@/components/main/mobile/common/MobileHeader";
+import MobileTabBar from "@/components/main/mobile/common/MobileTabBar";
 import MobileFooter from "@/components/main/mobile/common/MobileFooter";
 import DepartmentEditSheet from "@/components/main/adaptive/feature/MYP/DepartmentEditSheet";
 import DepartmentEditModal from "@/components/main/adaptive/feature/MYP/DepartmentEditModal";
